refactor: convert repo to monorepo structure
