+++ examples/react/index.ts
@@ -1,0 +1,9 @@
+export { AutoFormDemo } from "../autoForm/react/AutoFormDemo";
+export { RenderPropDemo } from "../autoForm/react/RenderPropDemo";
+export { ErrorComponentDemo } from "../autoForm/react/ErrorComponentDemo";
+export { NestedArrayDemo } from "../autoForm/react/NestedArrayDemo";
+export { SimpleNestedTest } from "../autoForm/react/SimpleNestedTest";
+export { ApiComparison } from "../autoForm/react/ApiComparison";
+export { UseFormDemo } from "../useForm/react/UseFormDemo";
+export { DirectArrayTest } from "../useForm/react/DirectArrayTest";
+
