+++ examples/useForm/react/DirectArrayTest.tsx
@@ -1,4 +1,4 @@
-import { useForm } from "../useForm";
+import { useForm } from "../../../packages/use-form/src";
 import { z } from "zod";
 
 // Test the useForm hook directly with array operations
