+++ package.json
@@ -1,31 +1,14 @@
 {
-  "name": "el-form",
-  "version": "0.1.0",
-  "description": "A minimal, headless form utility library for React + TypeScript",
-  "type": "module",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "name": "el-form-monorepo",
+  "private": true,
+  "packageManager": "pnpm@8.15.0",
+  "workspaces": ["packages/*"],
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
     "test": "echo \"No tests yet\" && exit 0"
   },
-  "keywords": [
-    "react",
-    "typescript",
-    "form",
-    "validation",
-    "zod",
-    "headless"
-  ],
-  "author": "",
-  "license": "MIT",
-  "packageManager": "pnpm@8.15.0",
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -41,9 +24,5 @@
     "tailwindcss": "^4.1.8",
     "typescript": "^5.2.2",
     "vite": "^5.0.8"
-  },
-  "peerDependencies": {
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0"
   }
 }
