+++ packages/auto-form/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@el-form/auto-form",
+  "version": "0.1.0",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": ["dist"],
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "peerDependencies": {
+    "react": ">=16.8.0",
+    "react-dom": ">=16.8.0"
+  },
+  "dependencies": {
+    "react": "^18.2.0",
+    "@el-form/use-form": "workspace:*"
+  }
+}
