+++ packages/auto-form/src/index.ts
@@ -1,0 +1,9 @@
+export { AutoForm } from "./AutoForm";
+export type {
+  AutoFormProps,
+  AutoFormFieldConfig,
+  AutoFormFieldProps,
+  AutoFormErrorProps,
+  GridColumns,
+  FieldConfig,
+} from "./types";
