+++ packages/auto-form/src/types.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
+import { UseFormReturn } from "../../use-form/src/types";
 
-// Grid system type for columns and spans (1-12)
 export type GridColumns = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
 
 export interface FieldConfig {
@@ -12,7 +12,6 @@ export interface FieldConfig {
   required?: boolean;
 }
 
-// Enhanced FieldConfig for nested arrays
 export interface AutoFormFieldConfig {
   name: string;
   label?: string;
@@ -28,48 +27,9 @@ export interface AutoFormFieldConfig {
   colSpan?: GridColumns;
   component?: React.ComponentType<AutoFormFieldProps>;
   options?: Array<{ value: string; label: string }>;
-  fields?: AutoFormFieldConfig[]; // For nested array fields
+  fields?: AutoFormFieldConfig[];
 }
 
-export interface FormState<T extends Record<string, any>> {
-  values: Partial<T>;
-  errors: Partial<Record<keyof T, string>>;
-  touched: Partial<Record<keyof T, boolean>>;
-  isSubmitting: boolean;
-  isValid: boolean;
-  isDirty: boolean;
-}
-
-export interface UseFormOptions<T extends Record<string, any>> {
-  schema: z.ZodSchema<T>;
-  initialValues?: Partial<T>;
-  onSubmit?: (values: T) => void | Promise<void>;
-  validateOnChange?: boolean;
-  validateOnBlur?: boolean;
-}
-
-export interface UseFormReturn<T extends Record<string, any>> {
-  register: <Name extends keyof T>(
-    name: Name
-  ) => {
-    name: Name;
-    value: T[Name];
-    onChange: (e: React.ChangeEvent<any>) => void;
-    onBlur: (e: React.FocusEvent<any>) => void;
-  };
-  handleSubmit: (
-    onValid: (data: T) => void,
-    onError?: (errors: Record<keyof T, string>) => void
-  ) => (e: React.FormEvent) => void;
-  formState: FormState<T>;
-  reset: () => void;
-  // Enhanced for nested arrays
-  setValue: (path: string, value: any) => void;
-  addArrayItem: (path: string, item: any) => void;
-  removeArrayItem: (path: string, index: number) => void;
-}
-
-// AutoForm Field Props for rendering
 export interface AutoFormFieldProps {
   name: string;
   label?: string;
@@ -81,7 +41,6 @@ export interface AutoFormFieldProps {
   error?: string;
   touched?: boolean;
   options?: Array<{ value: string; label: string }>;
-  // Array-specific props
   fields?: AutoFormFieldConfig[];
   onAddItem?: () => void;
   onRemoveItem?: (index: number) => void;
@@ -104,3 +63,4 @@ export interface AutoFormProps<T extends Record<string, any>> {
   children?: (formApi: UseFormReturn<T>) => React.ReactNode;
   customErrorComponent?: React.ComponentType<AutoFormErrorProps>;
 }
+
