+++ packages/use-form/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@el-form/use-form",
+  "version": "0.1.0",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": ["dist"],
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "peerDependencies": {
+    "react": ">=16.8.0"
+  },
+  "dependencies": {
+    "react": "^18.2.0",
+    "zod": "^3.22.4"
+  }
+}
