--- src/testApp/index.ts
@@ -1,15 +1,0 @@
-// Demo components for quick reference and comparison
-export { AutoFormDemo } from "./AutoFormDemo";
-export { UseFormDemo } from "./UseFormDemo";
-export { ApiComparison } from "./ApiComparison";
-
-// You can also create standalone demos by importing these individually:
-//
-// For a quick AutoForm example:
-// import { AutoFormDemo } from "./testApp/AutoFormDemo";
-//
-// For a useForm hook example:
-// import { UseFormDemo } from "./testApp/UseFormDemo";
-//
-// For API comparison info:
-// import { ApiComparison } from "./testApp/ApiComparison";
