+++ tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
+  content: ["./examples/**/*.{html,js,ts,jsx,tsx}", "./packages/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {},
   },
