+++ tsconfig.json
@@ -6,7 +6,6 @@
     "module": "ESNext",
     "skipLibCheck": true,
 
-    /* Bundler mode */
     "moduleResolution": "bundler",
     "allowImportingTsExtensions": true,
     "resolveJsonModule": true,
@@ -14,16 +13,14 @@
     "noEmit": true,
     "jsx": "react-jsx",
 
-    /* Linting */
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
 
-    /* Library build */
     "declaration": true,
     "outDir": "dist"
   },
-  "include": ["src"],
+  "include": ["packages/*/src", "src/index.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
